activity tracking added

+++ src/main/java/spring_server/chat_model/Message.java
@@ -22,5 +22,5 @@ public class Message {
 
     @Getter
     @Setter
-    private long chatRoomId;
+    private int chatRoomId;
 }
